:recycle: refactor(MonthlyPlanner): Split into reusable components.
